delete setMode.cpp and move them to encode.cpp
