+++ src/encode.h
@@ -2,3 +2,12 @@
    Message encoding
 */
 void encode(); // Loop through the string, transmitting one character at a time
+
+/*
+   TX modes
+*/
+
+void setModeJT9_1();
+void setModeJT9_2();
+void setModeWSPR();
+void setModeWSPR_telem();
