--- src/setMode.h
@@ -1,8 +1,0 @@
-/*
-   TX modes
-*/
-
-void setModeJT9_1();
-void setModeJT9_2();
-void setModeWSPR();
-void setModeWSPR_telem();
